Backend/app/requirements.txt
fastapi==0.110.0
psycopg2==2.9.7
asyncpg==0.29.0
SQLAlchemy==2.0.20
pydantic==2.5.3
uvicorn==0.27.1
httpx==0.24.1
passlib==1.7.4
bcrypt==3.1.7
PyJWT==2.7.0
python-dateutil==2.9.0
python-dotenv==1.0.1